+++ src/Numerics/LinearAlgebra/Complex32/Factorization/DenseLU.cs
@@ -64,7 +64,7 @@ namespace MathNet.Numerics.LinearAlgebra.Complex32.Factorization
             }
 
             // Create an array for the pivot indices.
-            var pivots = new int[matrix.RowCount];
+            var pivots = new long[matrix.RowCount];
 
             // Create a new matrix for the LU factors, then perform factorization (while overwriting).
             var factors = (DenseMatrix) matrix.Clone();
@@ -73,7 +73,7 @@ namespace MathNet.Numerics.LinearAlgebra.Complex32.Factorization
             return new DenseLU(factors, pivots);
         }
 
-        DenseLU(Matrix<Complex32> factors, int[] pivots)
+        DenseLU(Matrix<Complex32> factors, long[] pivots)
             : base(factors, pivots)
         {
         }
