+++ src/Numerics/LinearAlgebra/Complex32/Factorization/DenseLUBM.cs
@@ -65,7 +65,7 @@ namespace MathNet.Numerics.LinearAlgebra.Complex32.Factorization
             }
 
             // Create an array for the pivot indices.
-            var pivots = new int[matrix.RowCount];
+            var pivots = new long[matrix.RowCount];
 
             // Create a new matrix for the LU factors, then perform factorization (while overwriting).
             var factors = (DenseMatrixBM) matrix.Clone();
@@ -74,7 +74,7 @@ namespace MathNet.Numerics.LinearAlgebra.Complex32.Factorization
             return new DenseLUBM(factors, pivots);
         }
 
-        DenseLUBM(Matrix<Complex32> factors, int[] pivots)
+        DenseLUBM(Matrix<Complex32> factors, long[] pivots)
             : base(factors, pivots)
         {
         }
