+++ src/Numerics/LinearAlgebra/Single/Factorization/UserLU.cs
@@ -64,7 +64,7 @@ namespace MathNet.Numerics.LinearAlgebra.Single.Factorization
             // Create an array for the pivot indices.
             var order = matrix.RowCount;
             var factors = matrix.Clone();
-            var pivots = new int[order];
+            var pivots = new long[order];
 
             // Initialize the pivot matrix to the identity permutation.
             for (var i = 0; i < order; i++)
@@ -130,7 +130,7 @@ namespace MathNet.Numerics.LinearAlgebra.Single.Factorization
             return new UserLU(factors, pivots);
         }
 
-        UserLU(Matrix<float> factors, int[] pivots)
+        UserLU(Matrix<float> factors, long[] pivots)
             : base(factors, pivots)
         {
         }
@@ -181,8 +181,8 @@ namespace MathNet.Numerics.LinearAlgebra.Single.Factorization
                 var p = Pivots[i];
                 for (var j = 0; j < result.ColumnCount; j++)
                 {
-                    var temp = result.At(p, j);
-                    result.At(p, j, result.At(i, j));
+                    var temp = result.At((int)p, j);
+                    result.At((int)p, j, result.At(i, j));
                     result.At(i, j, temp);
                 }
             }
@@ -260,8 +260,8 @@ namespace MathNet.Numerics.LinearAlgebra.Single.Factorization
                 }
 
                 var p = Pivots[i];
-                var temp = result[p];
-                result[p] = result[i];
+                var temp = result[(int)p];
+                result[(int)p] = result[i];
                 result[i] = temp;
             }
