+++ src/Numerics/Providers/LinearAlgebra/Cuda/CudaLinearAlgebraProvider.Double.cs
@@ -214,7 +214,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra.Cuda
         /// <param name="ipiv">On exit, it contains the pivot indices. The size of the array must be <paramref name="order"/>.</param>
         /// <remarks>This is equivalent to the GETRF LAPACK routine.</remarks>
         [SecuritySafeCritical]
-        public override void LUFactor(double[] data, int order, int[] ipiv)
+        public override void LUFactor(double[] data, int order, long[] ipiv)
         {
             if (data == null)
             {
@@ -269,7 +269,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra.Cuda
         /// <param name="ipiv">The pivot indices of <paramref name="a"/>.</param>
         /// <remarks>This is equivalent to the GETRI LAPACK routine.</remarks>
         [SecuritySafeCritical]
-        public override void LUInverseFactored(double[] a, int order, int[] ipiv)
+        public override void LUInverseFactored(double[] a, int order, long[] ipiv)
         {
             if (a == null)
             {
@@ -338,7 +338,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra.Cuda
         /// <param name="b">On entry the B matrix; on exit the X matrix.</param>
         /// <remarks>This is equivalent to the GETRS LAPACK routine.</remarks>
         [SecuritySafeCritical]
-        public override void LUSolveFactored(int columnsOfB, double[] a, int order, int[] ipiv, double[] b)
+        public override void LUSolveFactored(int columnsOfB, double[] a, int order, long[] ipiv, double[] b)
         {
             if (a == null)
             {
