+++ src/Numerics/Providers/LinearAlgebra/ManagedLinearAlgebraProvider.Complex32.cs
@@ -1075,7 +1075,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra
         /// <param name="order">The order of the square matrix <paramref name="data"/>.</param>
         /// <param name="ipiv">On exit, it contains the pivot indices. The size of the array must be <paramref name="order"/>.</param>
         /// <remarks>This is equivalent to the GETRF LAPACK routine.</remarks>
-        public virtual void LUFactor(Complex32[] data, int order, int[] ipiv)
+        public virtual void LUFactor(Complex32[] data, int order, long[] ipiv)
         {
             if (data == null)
             {
@@ -1185,7 +1185,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra
                 throw new ArgumentException(Resources.ArgumentArraysSameLength, "a");
             }
 
-            var ipiv = new int[order];
+            var ipiv = new long[order];
             LUFactor(a, order, ipiv);
             LUInverseFactored(a, order, ipiv);
         }
@@ -1197,7 +1197,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra
         /// <param name="order">The order of the square matrix <paramref name="a"/>.</param>
         /// <param name="ipiv">The pivot indices of <paramref name="a"/>.</param>
         /// <remarks>This is equivalent to the GETRI LAPACK routine.</remarks>
-        public virtual void LUInverseFactored(Complex32[] a, int order, int[] ipiv)
+        public virtual void LUInverseFactored(Complex32[] a, int order, long[] ipiv)
         {
             if (a == null)
             {
@@ -1264,7 +1264,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra
                 throw new ArgumentException(Resources.ArgumentReferenceDifferent);
             }
 
-            var ipiv = new int[order];
+            var ipiv = new long[order];
             var clone = new Complex32[a.Length];
             a.Copy(clone);
             LUFactor(clone, order, ipiv);
@@ -1280,7 +1280,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra
         /// <param name="ipiv">The pivot indices of <paramref name="a"/>.</param>
         /// <param name="b">On entry the B matrix; on exit the X matrix.</param>
         /// <remarks>This is equivalent to the GETRS LAPACK routine.</remarks>
-        public virtual void LUSolveFactored(int columnsOfB, Complex32[] a, int order, int[] ipiv, Complex32[] b)
+        public virtual void LUSolveFactored(int columnsOfB, Complex32[] a, int order, long[] ipiv, Complex32[] b)
         {
             if (a == null)
             {
