+++ src/Numerics/Providers/LinearAlgebra/Mkl/MklLinearAlgebraProvider.Complex32.cs
@@ -251,7 +251,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra.Mkl
         /// <param name="ipiv">On exit, it contains the pivot indices. The size of the array must be <paramref name="order"/>.</param>
         /// <remarks>This is equivalent to the GETRF LAPACK routine.</remarks>
         [SecuritySafeCritical]
-        public override void LUFactor(Complex32[] data, int order, int[] ipiv)
+        public override void LUFactor(Complex32[] data, int order, long[] ipiv)
         {
             if (data == null)
             {
@@ -326,7 +326,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra.Mkl
         /// <param name="ipiv">The pivot indices of <paramref name="a"/>.</param>
         /// <remarks>This is equivalent to the GETRI LAPACK routine.</remarks>
         [SecuritySafeCritical]
-        public override void LUInverseFactored(Complex32[] a, int order, int[] ipiv)
+        public override void LUInverseFactored(Complex32[] a, int order, long[] ipiv)
         {
             if (a == null)
             {
@@ -415,7 +415,7 @@ namespace MathNet.Numerics.Providers.LinearAlgebra.Mkl
         /// <param name="b">On entry the B matrix; on exit the X matrix.</param>
         /// <remarks>This is equivalent to the GETRS LAPACK routine.</remarks>
         [SecuritySafeCritical]
-        public override void LUSolveFactored(int columnsOfB, Complex32[] a, int order, int[] ipiv, Complex32[] b)
+        public override void LUSolveFactored(int columnsOfB, Complex32[] a, int order, long[] ipiv, Complex32[] b)
         {
             if (a == null)
             {
